+++ NEWS.md
@@ -1,9 +1,10 @@
 
 # dev
 
-## Bug fixes
-
 * `cran_summary()` now works even if there is a NOTE/WARNING/ERROR in one platform and nothing on other platforms (@fabian-s, #259).
+
+* Check results have now a `get_ids()` method, to easily query the id(s)
+  of the check.
 
 # rhub 1.1.1